RobotDetection -> FeetDetection
