+++ crates/vision/src/lib.rs
@@ -1,11 +1,11 @@
 pub mod ball_detection;
 pub mod camera_matrix_extractor;
+pub mod feet_detection;
 pub mod field_border_detection;
 pub mod field_color_detection;
 pub mod image_receiver;
 pub mod image_segmenter;
 pub mod line_detection;
-pub mod luminance_image_extractor;
 pub mod perspective_grid_candidates_provider;
 mod ransac;
 pub mod robot_detection;
